+++ gamemodes/terrortown/gamemode/shared/sh_init.lua
@@ -523,12 +523,18 @@ require("marks")
 require("outline")
 
 include("includes/modules/pon.lua")
+
+-- include extensions
 include("ttt2/extensions/net.lua")
 include("ttt2/extensions/string.lua")
 include("ttt2/extensions/table.lua")
+include("ttt2/extensions/surface.lua")
 include("ttt2/extensions/draw.lua")
 
+-- include libraries
+include("ttt2/libraries/fonts.lua")
 include("ttt2/libraries/appearance.lua")
+include("ttt2/libraries/drawsc.lua")
 include("ttt2/libraries/vguihandler.lua")
 include("ttt2/libraries/vskin.lua")
 include("ttt2/libraries/fileloader.lua")